+++ tests/communication/tst_graphdatahandler.h
@@ -30,7 +30,7 @@ private slots:
 
 private:
 
-    void doHandleRegisterData(QList<Result<qint64> >& modbusResults, QList<QVariant> &actRawData);
+    void doHandleRegisterData(QList<Result<double> >& modbusResults, QList<QVariant> &actRawData);
 
     SettingsModel* _pSettingsModel;
     GraphDataModel* _pGraphDataModel;
